requirements.txt
numpy>=1.24
torch>=2.2
open_clip_torch>=2.24.0
Pillow>=10.0
pillow-heif>=0.15.0
streamlit>=1.32
tqdm>=4.66
